Update EloEstimationMethod.R
Removed P( Observation) base calculation outside of the for loop. this sped up the program from a 10 min run to a 6 minute run ish. also added back in the cycle program that runs a season multiple times.  decreases the error by very little for the 20 team setting on the test seeds 190022838, 190159096, and 000615600